fix(test): drop the redundant Proxy cast the type-aware lint rejects
new Proxy over a Float64Array is already typed; the cast tripped
no-unnecessary-type-assertion, which only runs in lint:type-aware.

+++ packages/core/tests/geometry-tile-resolution.test.ts
@@ -45,7 +45,7 @@ function counting(values: Float64Array): { array: Float64Array; reads: () => num
       if (typeof prop === "string" && /^\d+$/.test(prop)) reads += 1;
       return Reflect.get(target, prop) as unknown;
     },
-  }) as Float64Array;
+  });
   return { array, reads: () => reads };
 }
